refactor(core): remove unnecessary type cast in cycle accumulation
cpu_cycles is already u64, no need for as u64 cast when accumulating
into current_sample_cycles.

+++ fluxbench-core/src/bencher.rs
@@ -157,7 +157,7 @@ impl Bencher {
         } else {
             // During measurement: accumulate into current sample
             self.current_sample_time_ns += duration_nanos;
-            self.current_sample_cycles += cpu_cycles as u64;
+            self.current_sample_cycles += cpu_cycles;
             self.current_sample_iters += 1;
             self.current_sample_alloc_bytes += alloc_bytes;
             self.current_sample_alloc_count += alloc_count;
